working on setup params
