+++ .idea/misc.xml
@@ -31,6 +31,8 @@
         <entry key="app/src/main/res/layout/vh_payload.xml" value="0.3020833333333333" />
         <entry key="app/src/main/res/layout/vh_webman_option.xml" value="0.25298309326171864" />
         <entry key="app/src/main/res/layout/vh_webman_option_view.xml" value="0.75" />
+        <entry key="app/src/main/res/layout/view_ccapi.xml" value="0.27191107310816587" />
+        <entry key="app/src/main/res/layout/view_ps3mapi.xml" value="0.27191107310816587" />
         <entry key="app/src/main/res/layout/view_webman.xml" value="0.25" />
         <entry key="app/src/main/res/menu/bottom_menu.xml" value="0.3020833333333333" />
         <entry key="app/src/main/res/menu/ftp_long_click.xml" value="0.3020833333333333" />
